move security profile to cri/sputil pkg
