+++ internal/cri/server/container_create_linux_test.go
@@ -24,8 +24,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/containerd/containerd/v2/contrib/apparmor"
-	"github.com/containerd/containerd/v2/contrib/seccomp"
 	"github.com/containerd/containerd/v2/core/containers"
 	"github.com/containerd/containerd/v2/core/mount"
 	"github.com/containerd/containerd/v2/pkg/oci"
@@ -853,392 +851,6 @@ func TestNoDefaultRunMount(t *testing.T) {
 	}
 }
 
-func TestGenerateSeccompSecurityProfileSpecOpts(t *testing.T) {
-	for _, test := range []struct {
-		desc           string
-		profile        string
-		privileged     bool
-		disable        bool
-		specOpts       oci.SpecOpts
-		expectErr      bool
-		defaultProfile string
-		sp             *runtime.SecurityProfile
-	}{
-		{
-			desc:      "should return error if seccomp is specified when seccomp is not supported",
-			profile:   runtimeDefault,
-			disable:   true,
-			expectErr: true,
-		},
-		{
-			desc:    "should not return error if seccomp is not specified when seccomp is not supported",
-			profile: "",
-			disable: true,
-		},
-		{
-			desc:    "should not return error if seccomp is unconfined when seccomp is not supported",
-			profile: unconfinedProfile,
-			disable: true,
-		},
-		{
-			desc:       "should not set seccomp when privileged is true",
-			profile:    seccompDefaultProfile,
-			privileged: true,
-		},
-		{
-			desc:    "should not set seccomp when seccomp is unconfined",
-			profile: unconfinedProfile,
-		},
-		{
-			desc:    "should not set seccomp when seccomp is not specified",
-			profile: "",
-		},
-		{
-			desc:     "should set default seccomp when seccomp is runtime/default",
-			profile:  runtimeDefault,
-			specOpts: seccomp.WithDefaultProfile(),
-		},
-		{
-			desc:     "should set default seccomp when seccomp is docker/default",
-			profile:  dockerDefault,
-			specOpts: seccomp.WithDefaultProfile(),
-		},
-		{
-			desc:     "should set specified profile when local profile is specified",
-			profile:  profileNamePrefix + "test-profile",
-			specOpts: seccomp.WithProfile("test-profile"),
-		},
-		{
-			desc:           "should use default profile when seccomp is empty",
-			defaultProfile: profileNamePrefix + "test-profile",
-			specOpts:       seccomp.WithProfile("test-profile"),
-		},
-		{
-			desc:           "should fallback to docker/default when seccomp is empty and default is runtime/default",
-			defaultProfile: runtimeDefault,
-			specOpts:       seccomp.WithDefaultProfile(),
-		},
-		//-----------------------------------------------
-		// now buckets for the SecurityProfile variants
-		//-----------------------------------------------
-		{
-			desc:      "sp should return error if seccomp is specified when seccomp is not supported",
-			disable:   true,
-			expectErr: true,
-			sp: &runtime.SecurityProfile{
-				ProfileType: runtime.SecurityProfile_RuntimeDefault,
-			},
-		},
-		{
-			desc:    "sp should not return error if seccomp is unconfined when seccomp is not supported",
-			disable: true,
-			sp: &runtime.SecurityProfile{
-				ProfileType: runtime.SecurityProfile_Unconfined,
-			},
-		},
-		{
-			desc:       "sp should not set seccomp when privileged is true",
-			privileged: true,
-			sp: &runtime.SecurityProfile{
-				ProfileType: runtime.SecurityProfile_RuntimeDefault,
-			},
-		},
-		{
-			desc: "sp should not set seccomp when seccomp is unconfined",
-			sp: &runtime.SecurityProfile{
-				ProfileType: runtime.SecurityProfile_Unconfined,
-			},
-		},
-		{
-			desc: "sp should not set seccomp when seccomp is not specified",
-		},
-		{
-			desc:     "sp should set default seccomp when seccomp is runtime/default",
-			specOpts: seccomp.WithDefaultProfile(),
-			sp: &runtime.SecurityProfile{
-				ProfileType: runtime.SecurityProfile_RuntimeDefault,
-			},
-		},
-		{
-			desc:     "sp should set specified profile when local profile is specified",
-			specOpts: seccomp.WithProfile("test-profile"),
-			sp: &runtime.SecurityProfile{
-				ProfileType:  runtime.SecurityProfile_Localhost,
-				LocalhostRef: profileNamePrefix + "test-profile",
-			},
-		},
-		{
-			desc:     "sp should set specified profile when local profile is specified even without prefix",
-			specOpts: seccomp.WithProfile("test-profile"),
-			sp: &runtime.SecurityProfile{
-				ProfileType:  runtime.SecurityProfile_Localhost,
-				LocalhostRef: "test-profile",
-			},
-		},
-		{
-			desc:      "sp should return error if specified profile is invalid",
-			expectErr: true,
-			sp: &runtime.SecurityProfile{
-				ProfileType:  runtime.SecurityProfile_RuntimeDefault,
-				LocalhostRef: "test-profile",
-			},
-		},
-	} {
-		t.Run(test.desc, func(t *testing.T) {
-			cri := &criService{}
-			cri.config.UnsetSeccompProfile = test.defaultProfile
-			ssp := test.sp
-			csp, err := generateSeccompSecurityProfile(
-				test.profile,
-				test.defaultProfile)
-			if err != nil {
-				if test.expectErr {
-					assert.Error(t, err)
-				} else {
-					assert.NoError(t, err)
-				}
-			} else {
-				if ssp == nil {
-					ssp = csp
-				}
-				specOpts, err := cri.generateSeccompSpecOpts(ssp, test.privileged, !test.disable)
-				if test.expectErr {
-					assert.Error(t, err)
-				} else {
-					assert.NoError(t, err)
-					if test.specOpts == nil && specOpts == nil {
-						return
-					}
-					if test.specOpts == nil || specOpts == nil {
-						t.Fatalf("unexpected nil specOpts, expected nil: %t, actual nil: %t", test.specOpts == nil, specOpts == nil)
-					}
-					// `specOpts` for seccomp only uses/modifies `*specs.Spec`, not
-					// `oci.Client` or `*containers.Container`, so let's construct a
-					// `*specs.Spec` and compare if the results are the same.
-					expected := runtimespec.Spec{
-						Linux: &runtimespec.Linux{},
-						Process: &runtimespec.Process{
-							Capabilities: &runtimespec.LinuxCapabilities{
-								Bounding: []string{
-									"CAP_DAC_READ_SEARCH",
-									"CAP_SYS_ADMIN",
-									"CAP_SYS_BOOT",
-									"CAP_SYS_CHROOT",
-									"CAP_SYS_MODULE",
-									"CAP_SYS_PACCT",
-									"CAP_SYS_PTRACE",
-									"CAP_SYS_RAWIO",
-									"CAP_SYS_TIME",
-									"CAP_SYS_TTY_CONFIG",
-									"CAP_SYS_NICE",
-									"CAP_SYSLOG",
-									"CAP_BPF",
-									"CAP_PERFMON",
-								},
-							},
-						},
-					}
-					var actual runtimespec.Spec
-					err := util.DeepCopy(&actual, &expected)
-					assert.NoError(t, err)
-
-					test.specOpts(context.TODO(), nil, nil, &expected)
-					specOpts(context.TODO(), nil, nil, &actual)
-					assert.Equal(t, expected, actual)
-				}
-			}
-		})
-	}
-}
-
-func TestGenerateApparmorSpecOpts(t *testing.T) {
-	for _, test := range []struct {
-		desc       string
-		profile    string
-		privileged bool
-		disable    bool
-		specOpts   oci.SpecOpts
-		expectErr  bool
-		sp         *runtime.SecurityProfile
-	}{
-		{
-			desc:      "should return error if apparmor is specified when apparmor is not supported",
-			profile:   runtimeDefault,
-			disable:   true,
-			expectErr: true,
-		},
-		{
-			desc:    "should not return error if apparmor is not specified when apparmor is not supported",
-			profile: "",
-			disable: true,
-		},
-		{
-			desc:     "should set default apparmor when apparmor is not specified",
-			profile:  "",
-			specOpts: apparmor.WithDefaultProfile(appArmorDefaultProfileName),
-		},
-		{
-			desc:       "should not apparmor when apparmor is not specified and privileged is true",
-			profile:    "",
-			privileged: true,
-		},
-		{
-			desc:    "should not return error if apparmor is unconfined when apparmor is not supported",
-			profile: unconfinedProfile,
-			disable: true,
-		},
-		{
-			desc:    "should not apparmor when apparmor is unconfined",
-			profile: unconfinedProfile,
-		},
-		{
-			desc:       "should not apparmor when apparmor is unconfined and privileged is true",
-			profile:    unconfinedProfile,
-			privileged: true,
-		},
-		{
-			desc:     "should set default apparmor when apparmor is runtime/default",
-			profile:  runtimeDefault,
-			specOpts: apparmor.WithDefaultProfile(appArmorDefaultProfileName),
-		},
-		{
-			desc:       "should not apparmor when apparmor is default and privileged is true",
-			profile:    runtimeDefault,
-			privileged: true,
-		},
-		// TODO (mikebrow) add success with existing defined profile tests
-		{
-			desc:      "should return error when undefined local profile is specified",
-			profile:   profileNamePrefix + "test-profile",
-			expectErr: true,
-		},
-		{
-			desc:       "should return error when undefined local profile is specified and privileged is true",
-			profile:    profileNamePrefix + "test-profile",
-			privileged: true,
-			expectErr:  true,
-		},
-		{
-			desc:      "should return error if specified profile is invalid",
-			profile:   "test-profile",
-			expectErr: true,
-		},
-		//--------------------------------------
-		// buckets for SecurityProfile struct
-		//--------------------------------------
-		{
-			desc:      "sp should return error if apparmor is specified when apparmor is not supported",
-			disable:   true,
-			expectErr: true,
-			sp: &runtime.SecurityProfile{
-				ProfileType: runtime.SecurityProfile_RuntimeDefault,
-			},
-		},
-		{
-			desc:    "sp should not return error if apparmor is unconfined when apparmor is not supported",
-			disable: true,
-			sp: &runtime.SecurityProfile{
-				ProfileType: runtime.SecurityProfile_Unconfined,
-			},
-		},
-		{
-			desc: "sp should not apparmor when apparmor is unconfined",
-			sp: &runtime.SecurityProfile{
-				ProfileType: runtime.SecurityProfile_Unconfined,
-			},
-		},
-		{
-			desc:       "sp should not apparmor when apparmor is unconfined and privileged is true",
-			privileged: true,
-			sp: &runtime.SecurityProfile{
-				ProfileType: runtime.SecurityProfile_Unconfined,
-			},
-		},
-		{
-			desc:     "sp should set default apparmor when apparmor is runtime/default",
-			specOpts: apparmor.WithDefaultProfile(appArmorDefaultProfileName),
-			sp: &runtime.SecurityProfile{
-				ProfileType: runtime.SecurityProfile_RuntimeDefault,
-			},
-		},
-		{
-			desc:       "sp should not apparmor when apparmor is default and privileged is true",
-			privileged: true,
-			sp: &runtime.SecurityProfile{
-				ProfileType: runtime.SecurityProfile_RuntimeDefault,
-			},
-		},
-		{
-			desc:      "sp should return error when undefined local profile is specified",
-			expectErr: true,
-			sp: &runtime.SecurityProfile{
-				ProfileType:  runtime.SecurityProfile_Localhost,
-				LocalhostRef: profileNamePrefix + "test-profile",
-			},
-		},
-		{
-			desc:      "sp should return error when undefined local profile is specified even without prefix",
-			profile:   profileNamePrefix + "test-profile",
-			expectErr: true,
-			sp: &runtime.SecurityProfile{
-				ProfileType:  runtime.SecurityProfile_Localhost,
-				LocalhostRef: "test-profile",
-			},
-		},
-		{
-			desc:       "sp should return error when undefined local profile is specified and privileged is true",
-			privileged: true,
-			expectErr:  true,
-			sp: &runtime.SecurityProfile{
-				ProfileType:  runtime.SecurityProfile_Localhost,
-				LocalhostRef: profileNamePrefix + "test-profile",
-			},
-		},
-	} {
-		t.Run(test.desc, func(t *testing.T) {
-			asp := test.sp
-			csp, err := generateApparmorSecurityProfile(test.profile)
-			if err != nil {
-				if test.expectErr {
-					assert.Error(t, err)
-				} else {
-					assert.NoError(t, err)
-				}
-			} else {
-				if asp == nil {
-					asp = csp
-				}
-				specOpts, err := generateApparmorSpecOpts(asp, test.privileged, !test.disable)
-				if test.expectErr {
-					assert.Error(t, err)
-				} else {
-					assert.NoError(t, err)
-					if test.specOpts == nil && specOpts == nil {
-						return
-					}
-					if test.specOpts == nil || specOpts == nil {
-						t.Fatalf("unexpected nil specOpts, expected nil: %t, actual nil: %t", test.specOpts == nil, specOpts == nil)
-					}
-					// `specOpts` for seccomp only uses/modifies `*specs.Spec`, not
-					// `oci.Client` or `*containers.Container`, so let's construct a
-					// `*specs.Spec` and compare if the results are the same.
-					expected := runtimespec.Spec{
-						Linux:   &runtimespec.Linux{},
-						Process: &runtimespec.Process{},
-					}
-					var actual runtimespec.Spec
-					err := util.DeepCopy(&actual, &expected)
-					assert.NoError(t, err)
-
-					test.specOpts(context.TODO(), nil, nil, &expected)
-					specOpts(context.TODO(), nil, nil, &actual)
-					assert.Equal(t, expected, actual)
-				}
-			}
-		})
-	}
-}
-
 func TestMaskedAndReadonlyPaths(t *testing.T) {
 	testID := "test-id"
 	testSandboxID := "sandbox-id"
@@ -1381,77 +993,6 @@ func TestHostname(t *testing.T) {
 			require.NoError(t, err)
 			specCheck(t, testID, testSandboxID, testPid, spec)
 			assert.Contains(t, spec.Process.Env, test.expectedEnv)
-		})
-	}
-}
-
-func TestGenerateUserString(t *testing.T) {
-	type testcase struct {
-		// the name of the test case
-		name string
-
-		u        string
-		uid, gid *runtime.Int64Value
-
-		result        string
-		expectedError bool
-	}
-	testcases := []testcase{
-		{
-			name:   "Empty",
-			result: "",
-		},
-		{
-			name:   "Username Only",
-			u:      "testuser",
-			result: "testuser",
-		},
-		{
-			name:   "Username, UID",
-			u:      "testuser",
-			uid:    &runtime.Int64Value{Value: 1},
-			result: "testuser",
-		},
-		{
-			name:   "Username, UID, GID",
-			u:      "testuser",
-			uid:    &runtime.Int64Value{Value: 1},
-			gid:    &runtime.Int64Value{Value: 10},
-			result: "testuser:10",
-		},
-		{
-			name:   "Username, GID",
-			u:      "testuser",
-			gid:    &runtime.Int64Value{Value: 10},
-			result: "testuser:10",
-		},
-		{
-			name:   "UID only",
-			uid:    &runtime.Int64Value{Value: 1},
-			result: "1",
-		},
-		{
-			name:   "UID, GID",
-			uid:    &runtime.Int64Value{Value: 1},
-			gid:    &runtime.Int64Value{Value: 10},
-			result: "1:10",
-		},
-		{
-			name:          "GID only",
-			gid:           &runtime.Int64Value{Value: 10},
-			result:        "",
-			expectedError: true,
-		},
-	}
-	for _, tc := range testcases {
-		t.Run(tc.name, func(t *testing.T) {
-			r, err := generateUserString(tc.u, tc.uid, tc.gid)
-			if tc.expectedError {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-			}
-			assert.Equal(t, tc.result, r)
 		})
 	}
 }
